+++ src/futures/futures_unordered.rs
@@ -1,14 +1,9 @@
-// use futures::{
-//     stream::{FusedStream, Stream},
-//     task::AtomicWaker,
-// };
-
 use std::{
     cell::UnsafeCell,
     cmp,
     fmt::{self, Debug},
     future::Future,
-    iter::FromIterator,
+    marker::PhantomData,
     mem,
     pin::Pin,
     ptr,
@@ -22,7 +17,7 @@ use std::{
     task::{Context, Poll},
 };
 
-use super::{atomic_waker::AtomicWaker, Dequeue, ReadyToRunQueue, Task};
+use super::{atomic_waker::AtomicWaker, Dequeue, Iter, IterMut, IterPinMut, IterPinRef, ReadyToRunQueue, Task};
 
 /// Constant used for a `FuturesUnordered` to determine how many times it is
 /// allowed to poll underlying futures without yielding.
@@ -205,6 +200,53 @@ impl<Fut> FuturesUnordered<Fut> {
         // futures are ready. To do that we unconditionally enqueue it for
         // polling here.
         self.ready_to_run_queue.enqueue(ptr);
+    }
+
+    /// Returns an iterator that allows inspecting each future in the set.
+    pub fn iter(&self) -> Iter<'_, Fut>
+    where
+        Fut: Unpin,
+    {
+        Iter(Pin::new(self).iter_pin_ref())
+    }
+
+    /// Returns an iterator that allows inspecting each future in the set.
+    pub fn iter_pin_ref(self: Pin<&Self>) -> IterPinRef<'_, Fut> {
+        let (task, len) = self.atomic_load_head_and_len_all();
+        let pending_next_all = self.pending_next_all();
+
+        IterPinRef {
+            task,
+            len,
+            pending_next_all,
+            _marker: PhantomData,
+        }
+    }
+
+    /// Returns an iterator that allows modifying each future in the set.
+    pub fn iter_mut(&mut self) -> IterMut<'_, Fut>
+    where
+        Fut: Unpin,
+    {
+        IterMut(Pin::new(self).iter_pin_mut())
+    }
+
+    /// Returns an iterator that allows modifying each future in the set.
+    pub fn iter_pin_mut(mut self: Pin<&mut Self>) -> IterPinMut<'_, Fut> {
+        // `head_all` can be accessed directly and we don't need to spin on
+        // `Task::next_all` since we have exclusive access to the set.
+        let task = *self.head_all.get_mut();
+        let len = if task.is_null() {
+            0
+        } else {
+            unsafe { *(*task).len_all.get() }
+        };
+
+        IterPinMut {
+            task,
+            len,
+            _marker: PhantomData,
+        }
     }
 
     /// Returns the current head node and number of futures in the list of all
@@ -561,18 +603,5 @@ impl<Fut> Drop for FuturesUnordered<Fut> {
         // While that freeing operation isn't guaranteed to happen here, it's
         // guaranteed to happen "promptly" as no more "blocking work" will
         // happen while there's a strong refcount held.
-    }
-}
-
-impl<Fut> FromIterator<Fut> for FuturesUnordered<Fut> {
-    fn from_iter<I>(iter: I) -> Self
-    where
-        I: IntoIterator<Item = Fut>,
-    {
-        let acc = Self::new();
-        iter.into_iter().fold(acc, |acc, item| {
-            acc.push(item);
-            acc
-        })
     }
 }