+++ src/futures/mod.rs
@@ -62,11 +62,11 @@ pub trait ArcWake: Send + Sync {
 pub mod enter;
 
 mod atomic_waker;
+mod iter;
 
 // mod iter;
 use std::sync::Arc;
-
-// pub use self::iter::{IntoIter, Iter, IterMut, IterPinMut, IterPinRef};
+pub use self::iter::{IntoIter, Iter, IterMut, IterPinMut, IterPinRef};
 
 mod futures_unordered;
 pub use self::futures_unordered::FuturesUnordered;